[#14] AnalyzeView에 Object Detection 이미지 전달
