+++ Junction2024-Team9/Junction2024-Team9/App/Junction2024_Team9App.swift
@@ -10,21 +10,16 @@ import SwiftUI
 @main
 struct Junction2024_Team9App: App {
     @UIApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
-    @StateObject private var appRouter: AppRouter = .init()
     @StateObject private var locationHelper = LocationHelper.shared
     
     var body: some Scene {
 
         WindowGroup {
-            NavigationStack(path: $appRouter.path) {
-                HomeView()
-                    .navigationDestination(for: AppRoute.self) { route in
-                        appRouter.navDestVB(to: route)
-                    }
+            NavigationStack() {
+                CaptureView()
             }
             
         }
-        .environmentObject(appRouter)
         .environmentObject(locationHelper)
     }
 }
